addon server command - to simulate a full Bitrise Add-on Login flow
So far the bitrise-addon-test toolkit was only able to send and test the login call
but when you're working with an addon where the Frontend and Backend
are two separate services, the Addon Login request
goes to the Backend which generates a JWT token, stores it in a cookie
and then redirects to the Frontend, this flow was super complicated to test.

With this new "login server" command you can now use the testkit to spin up a minimal
webserver so you can open the "login" page in a web browser
and click the "login" button to simulate the whole
login flow, not just to validate that the your addon responds
with the required response to the login request.